+++ platform.js
@@ -14,6 +14,7 @@ class App {
   constructor (opts, templatePath) {
     this.opts = opts
     this.templatePath = templatePath
+    this.asarOptions = common.createAsarOpts(opts)
 
     if (this.opts.prune === undefined) {
       this.opts.prune = true
@@ -68,6 +69,10 @@ class App {
     }
   }
 
+  get appAsarPath () {
+    return path.join(this.originalResourcesDir, 'app.asar')
+  }
+
   relativeRename (basePath, oldName, newName) {
     debug(`Renaming ${oldName} to ${newName} in ${basePath}`)
     return fs.rename(path.join(basePath, oldName), path.join(basePath, newName))
@@ -80,11 +85,14 @@ class App {
   /**
    * Performs the following initial operations for an app:
    * * Creates temporary directory
-   * * Copies template into temporary directory
-   * * Copies user's app into temporary directory
-   * * Prunes non-production node_modules (if opts.prune is either truthy or undefined)
    * * Remove default_app (which is either a folder or an asar file)
-   * * Creates an asar (if opts.asar is set)
+   * * If a prebuilt asar is specified:
+   *   * Copies asar into temporary directory as app.asar
+   * * Otherwise:
+   *   * Copies template into temporary directory
+   *   * Copies user's app into temporary directory
+   *   * Prunes non-production node_modules (if opts.prune is either truthy or undefined)
+   *   * Creates an asar (if opts.asar is set)
    *
    * Prune and asar are performed before platform-specific logic, primarily so that
    * this.originalResourcesAppDir is predictable (e.g. before .app is renamed for Mac)
@@ -93,32 +101,40 @@ class App {
     debug(`Initializing app in ${this.stagingPath} from ${this.templatePath} template`)
 
     return fs.move(this.templatePath, this.stagingPath, { clobber: true })
-      .then(() => this.copyTemplate())
       .then(() => this.removeDefaultApp())
+      .then(() => {
+        if (this.opts.prebuiltAsar) {
+          return this.copyPrebuiltAsar()
+        } else {
+          return this.buildApp()
+        }
+      })
+  }
+
+  buildApp () {
+    return this.copyTemplate()
       .then(() => this.asarApp())
   }
 
   copyTemplate () {
-    return fs.copy(this.opts.dir, this.originalResourcesAppDir, {
-      filter: ignore.userIgnoreFilter(this.opts),
-      dereference: this.opts.derefSymlinks
-    }).then(() => hooks.promisifyHooks(this.opts.afterCopy, [
+    const hookArgs = [
       this.originalResourcesAppDir,
       this.opts.electronVersion,
       this.opts.platform,
       this.opts.arch
-    ])).then(() => {
-      if (this.opts.prune) {
-        return hooks.promisifyHooks(this.opts.afterPrune, [
-          this.originalResourcesAppDir,
-          this.opts.electronVersion,
-          this.opts.platform,
-          this.opts.arch
-        ])
-      } else {
-        return true
-      }
+    ]
+
+    return fs.copy(this.opts.dir, this.originalResourcesAppDir, {
+      filter: ignore.userIgnoreFilter(this.opts),
+      dereference: this.opts.derefSymlinks
     })
+      .then(() => hooks.promisifyHooks(this.opts.afterCopy, hookArgs))
+      .then(() => {
+        if (this.opts.prune) {
+          return hooks.promisifyHooks(this.opts.afterPrune, hookArgs)
+        }
+        return true
+      })
   }
 
   removeDefaultApp () {
@@ -146,15 +162,47 @@ class App {
       .catch(/* istanbul ignore next */ () => null)
   }
 
+  prebuiltAsarWarning (option, triggerWarning) {
+    if (triggerWarning) {
+      common.warning(`prebuiltAsar and ${option} are incompatible, ignoring the ${option} option`)
+    }
+  }
+
+  copyPrebuiltAsar () {
+    if (this.asarOptions) {
+      common.warning('prebuiltAsar has been specified, all asar options will be ignored')
+    }
+
+    for (const hookName of ['afterCopy', 'afterPrune']) {
+      if (this.opts[hookName]) {
+        throw new Error(`${hookName} is incompatible with prebuiltAsar`)
+      }
+    }
+
+    this.prebuiltAsarWarning('ignore', this.opts.originalIgnore)
+    this.prebuiltAsarWarning('prune', !this.opts.prune)
+    this.prebuiltAsarWarning('derefSymlinks', this.opts.derefSymlinks !== undefined)
+
+    const src = path.resolve(this.opts.prebuiltAsar)
+
+    return fs.stat(src)
+      .then(stat => {
+        if (!stat.isFile()) {
+          throw new Error(`${src} specified in prebuiltAsar must be an asar file.`)
+        }
+
+        debug(`Copying asar: ${src} to ${this.appAsarPath}`)
+        return fs.copy(src, this.appAsarPath, {overwrite: false, errorOnExist: true})
+      })
+  }
+
   asarApp () {
-    const asarOptions = common.createAsarOpts(this.opts)
-    if (!asarOptions) {
+    if (!this.asarOptions) {
       return Promise.resolve()
     }
 
-    const dest = path.join(this.originalResourcesDir, 'app.asar')
-    debug(`Running asar with the options ${JSON.stringify(asarOptions)}`)
-    return pify(asar.createPackageWithOptions)(this.originalResourcesAppDir, dest, asarOptions)
+    debug(`Running asar with the options ${JSON.stringify(this.asarOptions)}`)
+    return pify(asar.createPackageWithOptions)(this.originalResourcesAppDir, this.appAsarPath, this.asarOptions)
       .then(() => fs.remove(this.originalResourcesAppDir))
   }