+++ test/asar.js
@@ -3,6 +3,7 @@
 const common = require('../common')
 const path = require('path')
 const test = require('ava')
+const sinon = require('sinon')
 const util = require('./_util')
 
 test('asar argument test: asar is not set', t => {
@@ -49,3 +50,47 @@ util.testSinglePlatform('asar test', (t, opts) => {
       ])
     })
 })
+
+util.testSinglePlatform('prebuilt asar test', (t, opts) => {
+  opts.name = 'prebuiltAsarTest'
+  opts.dir = util.fixtureSubdir('asar-prebuilt')
+  opts.prebuiltAsar = path.join(opts.dir, 'app.asar')
+  opts.asar = {
+    'unpack': '*.pac',
+    'unpackDir': 'dir_to_unpack'
+  }
+  opts.ignore = ['foo']
+  opts.prune = false
+  opts.derefSymlinks = false
+  sinon.spy(console, 'warn')
+
+  let resourcesPath
+  return util.packageAndEnsureResourcesPath(t, opts)
+    .then(generatedResourcesPath => {
+      util.assertWarning(t, 'WARNING: prebuiltAsar has been specified, all asar options will be ignored')
+      for (const incompatibleOption of ['ignore', 'prune', 'derefSymlinks']) {
+        util.assertWarning(t, `WARNING: prebuiltAsar and ${incompatibleOption} are incompatible, ignoring the ${incompatibleOption} option`)
+      }
+
+      resourcesPath = generatedResourcesPath
+      return util.assertFile(t, path.join(resourcesPath, 'app.asar'), 'app.asar should exist under the resources subdirectory when opts.prebuiltAsar points to a prebuilt asar')
+    }).then(() => util.assertFilesEqual(t, opts.prebuiltAsar, path.join(resourcesPath, 'app.asar'), 'app.asar should equal the prebuilt asar'))
+    .then(() => util.assertPathNotExists(t, path.join(resourcesPath, 'app'), 'app subdirectory should NOT exist when app.asar is built'))
+})
+
+function testFailedPrebuiltAsar (name, extraOpts, errorRegex) {
+  const dir = util.fixtureSubdir('asar-prebuilt')
+  util.testSinglePlatform(`prebuilt asar: fail on ${name}`, util.invalidOptionTest(Object.assign({
+    name: 'prebuiltAsarFailingTest',
+    dir: dir,
+    prebuiltAsar: path.join(dir, 'app.asar')
+  }, extraOpts), errorRegex))
+}
+
+function testIncompatibleOptionWithPrebuiltAsar (extraOpts) {
+  testFailedPrebuiltAsar(`specifying prebuiltAsar and ${Object.keys(extraOpts).join(',')}`, extraOpts, /is incompatible with prebuiltAsar/)
+}
+
+testFailedPrebuiltAsar('prebuiltAsar set to directory', { prebuiltAsar: util.fixtureSubdir('asar-prebuilt') }, /must be an asar file/)
+testIncompatibleOptionWithPrebuiltAsar({ afterCopy: [] })
+testIncompatibleOptionWithPrebuiltAsar({ afterPrune: [] })